[Developer] Refactor index.ts and cleanup package.json
