+++ developer/js/tests/test-compile-trie.ts
@@ -1,4 +1,4 @@
-import LexicalModelCompiler from '../';
+import LexicalModelCompiler from '../dist/lexical-model-compiler/lexical-model-compiler';
 import {assert} from 'chai';
 import 'mocha';
