+++ developer/js/tests/test-punctuation.ts
@@ -1,4 +1,4 @@
-import LexicalModelCompiler from '../';
+import LexicalModelCompiler from '../dist/lexical-model-compiler/lexical-model-compiler';
 import {assert} from 'chai';
 import 'mocha';
 
@@ -7,7 +7,7 @@ import { compileModelSourceCode } from './helpers';
 
 
 describe('LexicalModelCompiler', function () {
-  describe('spefifying punctuation', function () {
+  describe('specifying punctuation', function () {
     const MODEL_ID = 'example.qaa.trivial';
     const PATH = path.join(__dirname, 'fixtures', MODEL_ID)
